third_party/blink/web_tests/external/wpt/html/links/stylesheet/quirk-origin-check-expected.txt

This is a testharness.js-based test.
[FAIL] Origin check for stylesheet with non-CSS MIME type quirk
  Uncaught Error: assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
